+++ public/index.html
@@ -46,6 +46,6 @@
 
       <!-- Контент -->
     </main>
-    <script src = "../src/main.js"></script>
+    <script type="module" src = "../src/main.js"></script>
   </body>
 </html>
