+++ src/components/load-more.js
@@ -1,0 +1,6 @@
+
+export const createLoadMoreTemplate = () => {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
